Rough attempt to fix class does not have a __table__ or  __tablename__ specified and does not inherit from an existing table-mapped class exception
This is being triggered when configure_supporting_documents_inheritance
is being called with profile_supporting_docs_t having a None value

It appears as the base class isn't actually being used anywhere and
can be safely removed, as the EntitySupportingDocumentProxy subclass
overwrites the __table__ and __mapper__ values set in the base class
anyway.

The one concern is how to correctly handle p_doc_id_col when profile_supporting_docs_t
is None. Here I've also set that to None, but I suspect that will cause
issues with the

    'inherit_condition': t_doc_id_col == p_doc_id_col

part of EntitySupportingDocumentProxy.