DocumentGenerator: Fix format of generated SQL
This fixes the issue of properly formatting the SQL produced when
generating a document. To produce the correct SQL, the name of the
view must have the following format: 'viewname_vw_someothername',
"_vw_" *MUST* be in the name

However, to conform to STDM table naming style, we recommend to start
the view name with the profile prefix e.g.

"ka_analysis_vw_persondata"