Composer: Add a QGIS2 to QGIS3 template converter
Due to the changes in the composer design in QGIS3, it is impossible
to open old QGIS2 templates designed with STDM composer items in QGIS3

The purpose of this utility is to automatically read all your
existing QGIS2 templates (from your default STDM template folder)
and convert them to QGIS3 format. Upon a successful conversion,
the utility creates a new file with the same name as the old one,
but with an appended prefix "_v3".

STDM composer items converted by this utility are:
- DataLabel item
- Photo item
- Chart item
- Table item
- Map item
- QRCode item