Data Import: Check for machine date format
Previously, when you imported data (csv) that contained a date
column, you had to make sure that the date format in the column
matches the machine date format, without this you would get a stack
trace during the importation process.

This patch checks this date format automatically and report any mis-match
with a user-friendly message. However,this relies on the format saved
in Windows registry hence the patch will only work for STDM running
on Windows OS.